**Very very bad implementations**

+++ src/main/java/app/repo/MeatRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.Meats;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MeatRepo extends CrudRepository<Meats,Long> {
+public interface MeatRepo extends CrudRepository<Meats,Integer> {
 }
