+++ src/main/java/app/repo/OrderRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.PizzaOrder;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepo extends CrudRepository<PizzaOrder, Long> {
+public interface OrderRepo extends CrudRepository<PizzaOrder, Integer> {
 }
