+++ src/main/java/app/repo/SauceRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.Sauces;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SauceRepo extends CrudRepository<Sauces,Long> {
+public interface SauceRepo extends CrudRepository<Sauces,Integer> {
 }
