+++ src/main/java/app/repo/VeggiesRepo.java
@@ -2,7 +2,6 @@ package app.repo;
 
 import app.entity.Veggies;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
-public interface VeggiesRepo extends CrudRepository<Veggies,Long> {
+public interface VeggiesRepo extends CrudRepository<Veggies,Integer> {
 }
